Nation Post V1.4/NationPost/backend/requirements.txt
flask
flask-cors
pymongo
bcrypt
dnspython
uuid
requests
google-generativeai
PyJWT